+++ Token.h
@@ -3,7 +3,7 @@ struct Token {
 	char* symbols;
 	short int arguments = 0; // amount of operands
 	short int index = 0;
-	size_t size = 0;
+	short int size = 0;
 	Token* next = nullptr;
 	Token* prev = nullptr;
 	Token& operator=(Token&& t);
